unmodded_DCRNN_PyTorch/requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
anyio==4.8.0
astunparse==1.6.3
attrs==24.3.0
blosc2==3.3.0
bokeh==3.4.2
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.7
cloudpickle==3.0.0
contourpy==1.2.1
Cython==3.0.11
dask==2024.4.2
dask-expr==1.0.14
dask-pytorch-ddp==0.2.2
decorator==4.4.2
distributed==2024.4.2
filelock==3.13.1
flatbuffers==25.1.24
frozenlist==1.5.0
fsspec==2024.2.0
gast==0.6.0
google-pasta==0.2.0
grpcio==1.70.0
h11==0.14.0
h5py==3.12.1
httpcore==1.0.7
httpx==0.28.1
idna==3.10
importlib_metadata==8.0.0
Jinja2==3.1.3
keras==3.8.0
libclang==18.1.1
locket==1.0.0
lz4==4.3.3
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
ml-dtypes==0.4.1
mpmath==1.2.1
msgpack==1.0.8
multidict==6.1.0
namex==0.0.8
ndindex==1.9.2
networkx==3.2.1
numexpr==2.10.2
numpy==1.26.4
nvidia-cublas-cu12==12.4.2.65
nvidia-cuda-cupti-cu12==12.4.99
nvidia-cuda-nvrtc-cu12==12.4.99
nvidia-cuda-runtime-cu12==12.4.99
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.0.44
nvidia-curand-cu12==10.3.5.119
nvidia-cusolver-cu12==11.6.0.99
nvidia-cusparse-cu12==12.3.0.142
nvidia-ml-py==12.560.30
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.4.99
nvidia-nvtx-cu12==12.4.99
opt_einsum==3.4.0
optree==0.14.0
packaging==24.1
pandas==2.2.2
partd==1.4.2
Pillow==9.3.0
propcache==0.2.1
protobuf==5.29.3
psutil==6.0.0
py-cpuinfo==9.0.0
pyarrow==16.1.0
pyarrow-hotfix==0.6
Pygments==2.19.1
pynvml==12.0.0
pyparsing==3.2.1
python-dateutil==2.9.0.post0
pytorch-triton==3.0.0+dedb7bdf33
pytz==2024.1
PyYAML==6.0.1
requests==2.32.3
rich==13.9.4
scipy==1.15.1
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
sympy==1.12
tables==3.10.2
tblib==3.0.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==2.5.0
toolz==0.12.1
torch==2.5.0.dev20240627+cu124
torch_scatter==2.1.2
torch_sparse==0.6.18
torchaudio==2.4.0.dev20240627+cu124
torchvision==0.20.0.dev20240627+cu124
tornado==6.4.1
tqdm==4.67.1
typing_extensions==4.8.0
tzdata==2024.1
urllib3==2.3.0
utils==1.0.2
Werkzeug==3.1.3
wrapt==1.17.2
xyzservices==2024.6.0
yarl==1.18.3
zict==3.0.0
zipp==3.19.2
